feature/backend/add auth structure + change authRoutes/LIN
